pipelines/scripts/requirements.txt
azure-ai-ml
azure-ai-ml
azure-identity
azure-storage-blob
azure-storage-file-datalake
azureml-dataprep
azureml-dataprep-native
azureml-dataprep-rslex
azureml-fsspec
fsspec
joblib
mltable
numpy
pandas
pyarrow
python-dotenv
python-dateutil
PyYAML
scikit-learn
scipy
strictyaml
tqdm
setuptools #>=71.0.2,<80.9
wheel
fastparquet
matplotlib
seaborn